% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km2mil.R
\name{km2mil}
\alias{km2mil}
\title{Kilometres to Miles}
\usage{
km2mil(kilometres)
}
\arguments{
\item{kilometres}{A numeric value or numeric vector}
}
\value{
the Distance in Miles
}
\description{
Conversion the Distance in Kilometres to Miles
}
\examples{
dist1 <- km2mil(71)
dist2 <- km2mil(c(16, 71, 88))

}
