% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mil2km.R
\name{mil2km}
\alias{mil2km}
\title{Miles to Kilometres}
\usage{
mil2km(miles)
}
\arguments{
\item{miles}{A numeric value or numeric vector}
}
\value{
the distance in kilometres
}
\description{
Conversion the distance in miles to kilometres
}
\examples{
dist1 <- mil2km(71)
dist2 <- mil2km(c(16, 71, 88))

}
